+++ src/main/java/com/zivy009/demo/springbootshirodwz/service/impl/SysRoleService.java
@@ -151,7 +151,7 @@ public class SysRoleService extends ServiceImpl<SysRoleMapper, SysRole> implemen
      * @return
      *
      */
-    public Set<Integer> obtainRolesSelf(Integer userId) {
+    public Set<Integer> obtainRolesSelf(Long userId) {
 
         return sysUserRoleMapper.obtainRoldidBy(userId);
